use serde and bincode for column metadata serialization and deserialization
