+++ src/node/src/lib.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-extern crate core;
+extern crate kernel;
 extern crate log;
 extern crate protocol;
 extern crate storage;
