+++ src/sql_types/Cargo.toml
@@ -9,3 +9,4 @@ publish = false
 kernel = { path = "../kernel" }
 lexical = "5.2.0"
 rstest = "0.6.4"
+serde = { version = "1.0.114", features = ["derive"] }
